Extract stats report shared helpers
